Replace 8-bit parent keys with 16-bit dual-fold multi-posting.
Exact parent lookup via stable fold selection replaces HD-1 expansion.

+++ include/vectorcache/query/engine.hpp
@@ -23,7 +23,7 @@ struct QueryParams {
 
 struct PreparedQuery {
   AlignedVector<float> rotated;
-  std::uint8_t parent_key = 0;
+  std::uint16_t parent_key = 0;
   AlignedVector<std::uint64_t> l0;
 };
